# docs/swagger.yaml
definitions:
  handler.EmailRequest:
    properties:
      email:
        example: johndoe@example.com
        type: string
    required:
    - email
    type: object
  handler.RegisterForm:
    properties:
      email:
        example: johndoe@example.com
        type: string
      password:
        example: secret123
        type: string
      username:
        example: johndoe
        type: string
    required:
    - email
    - password
    - username
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  response.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
    type: object
info:
  contact: { }
  version: 1.0.0
  title: ""
paths:
  /healthcheck:
    get:
      consumes:
      - application/json
      description: Returns a simple message to check if the API is alive
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
      summary: Show the status of the server
      tags:
      - root
  /users/find:
    post:
      consumes:
      - application/json
      description: Retrieves a user record using their email address.
      parameters:
      - description: Email search request
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/handler.EmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Find user by email
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account by providing username, password, and
        email.
      parameters:
      - description: User registration form
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterForm'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register a new user
      tags:
      - users
swagger: "2.0"